+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/allocation/FirstFitAllocator.java
@@ -176,6 +176,6 @@ public class FirstFitAllocator extends OrderedAllocator {
       allocateMemoryObject(vertex, firstFitOffset);
     }
 
-    return getMemorySize();
+    return getMemorySizeInByte();
   }
 }