+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/allocation/tasks/MemoryScriptTask.java
@@ -40,7 +40,6 @@ import bsh.EvalError;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.emf.common.util.EMap;
 import org.preesm.algorithm.memory.exclusiongraph.MemoryExclusionGraph;
 import org.preesm.algorithm.memory.script.MemoryScriptEngine;
 import org.preesm.algorithm.model.dag.DirectedAcyclicGraph;
@@ -50,6 +49,7 @@ import org.preesm.commons.doc.annotations.PreesmTask;
 import org.preesm.commons.doc.annotations.Value;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.model.scenario.Scenario;
+import org.preesm.model.scenario.SimulationInfo;
 import org.preesm.model.scenario.check.FifoTypeChecker;
 import org.preesm.workflow.elements.Workflow;
 import org.preesm.workflow.implement.AbstractTaskImplementation;
@@ -84,6 +84,12 @@ import org.preesm.workflow.implement.AbstractTaskImplementation;
                 @Value(name = "None",
                     effect = "No verification is performed. Use this policy to speed up workflow execution once all"
                         + " memory scripts have been validated..") }),
+        @Parameter(name = "False Sharing Prevention",
+            description = "Force additional allocation before/after buffer to prevent false sharing issues.",
+            values = { @Value(name = "False", effect = "The false sharing prevention mecanism will not be used."),
+                @Value(name = "True",
+                    effect = "The false sharing prevention mecanism will be used."
+                        + "Using the Data alignement parameter.") }),
         @Parameter(name = "Data alignment",
             description = "Option used to force the allocation of buffers with aligned addresses. The data"
                 + " alignment property should always have the same value as the one set in the properties of "
@@ -91,10 +97,10 @@ import org.preesm.workflow.implement.AbstractTaskImplementation;
             values = { @Value(name = "None", effect = "No special care is taken to align the buffers in memory."),
                 @Value(name = "Data",
                     effect = "All buffers are aligned on addresses that are multiples of their size. For example,"
-                        + "a 4 bytes integer is aligned on 4 bytes address."),
+                        + "a 32 bits integer is aligned on 32 bits address."),
                 @Value(name = "Fixed:=$$n$$",
                     effect = "Where $$n\\in \\mathbb{N}^*$$. This forces the allocation algorithm to align all buffers"
-                        + " on addresses that are multiples of n bytes.") }),
+                        + " on addresses that are multiples of n bits.") }),
         @Parameter(name = "Log Path",
             description = "Specify whether, and where, a log of the buffer matching optimization should be "
                 + "generated. Generated log are in the markdown format, and provide information "
@@ -123,6 +129,7 @@ public class MemoryScriptTask extends AbstractTaskImplementation {
   public static final String VALUE_CHECK_NONE     = "None";
   public static final String VALUE_CHECK_FAST     = "Fast";
   public static final String VALUE_CHECK_THOROUGH = "Thorough";
+  public static final String PARAM_FALSE_SHARING  = "False Sharing Prevention";
 
   @Override
   public Map<String, String> getDefaultParameters() {
@@ -131,6 +138,7 @@ public class MemoryScriptTask extends AbstractTaskImplementation {
         "? C {" + MemoryScriptTask.VALUE_TRUE + ", " + MemoryScriptTask.VALUE_FALSE + "}");
     param.put(MemoryScriptTask.PARAM_CHECK, "? C {" + MemoryScriptTask.VALUE_CHECK_NONE + ", "
         + MemoryScriptTask.VALUE_CHECK_FAST + ", " + MemoryScriptTask.VALUE_CHECK_THOROUGH + "}");
+    param.put(MemoryScriptTask.PARAM_FALSE_SHARING, MemoryScriptTask.VALUE_FALSE);
     param.put(MemoryAllocatorTask.PARAM_ALIGNMENT, MemoryAllocatorTask.VALUE_ALIGNEMENT_DEFAULT);
     param.put(MemoryScriptTask.PARAM_LOG, MemoryScriptTask.VALUE_LOG);
 
@@ -152,6 +160,11 @@ public class MemoryScriptTask extends AbstractTaskImplementation {
     // Get the log parameter
     final String log = parameters.get(MemoryScriptTask.PARAM_LOG);
 
+    // Get false sharing prevention flag
+    boolean falseSharingPreventionFlag = false;
+    falseSharingPreventionFlag = parameters.get(MemoryScriptTask.PARAM_FALSE_SHARING)
+        .equals(MemoryScriptTask.VALUE_TRUE);
+
     // Retrieve the alignment param
     final String valueAlignment = parameters.get(MemoryAllocatorTask.PARAM_ALIGNMENT);
 
@@ -161,7 +174,7 @@ public class MemoryScriptTask extends AbstractTaskImplementation {
     // Get the data types from the scenario
     final Scenario scenario = (Scenario) inputs.get("scenario");
     FifoTypeChecker.checkMissingFifoTypeSizes(scenario);
-    final EMap<String, Long> dataTypes = scenario.getSimulationInfo().getDataTypes();
+    final SimulationInfo simulationInfo = scenario.getSimulationInfo();
 
     // Get check policy
     final String checkString = parameters.get(MemoryScriptTask.PARAM_CHECK);
@@ -169,9 +182,9 @@ public class MemoryScriptTask extends AbstractTaskImplementation {
     final MemoryExclusionGraph meg = (MemoryExclusionGraph) inputs.get("MemEx");
 
     // execute
-    final MemoryScriptEngine engine = new MemoryScriptEngine(valueAlignment, log, verbose);
+    final MemoryScriptEngine engine = new MemoryScriptEngine(falseSharingPreventionFlag, valueAlignment, log, verbose);
     try {
-      engine.runScripts(dag, dataTypes, checkString);
+      engine.runScripts(dag, simulationInfo, checkString);
     } catch (final EvalError e) {
       final String message = "An error occurred during memory scripts execution";
       throw new PreesmRuntimeException(message, e);