+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/script/Match.java
@@ -139,6 +139,10 @@ public class Match {
     return this.localIndex;
   }
 
+  public long getLocalIndexInByte() {
+    return (this.localIndex + 7L) / 8L;
+  }
+
   public void setLocalIndex(final long localIndex) {
     this.localIndex = localIndex;
   }
@@ -155,12 +159,20 @@ public class Match {
     return this.remoteIndex;
   }
 
+  public long getRemoteIndexInByte() {
+    return (this.remoteIndex + 7L) / 8L;
+  }
+
   public void setRemoteIndex(final long remoteIndex) {
     this.remoteIndex = remoteIndex;
   }
 
   public long getLength() {
     return this.length;
+  }
+
+  public long getLengthInByte() {
+    return (this.length + 7L) / 8L;
   }
 
   public void setLength(final long length) {
@@ -333,9 +345,10 @@ public class Match {
 
   @Override
   public String toString() {
-    return "" + getLocalBuffer().getVertexName() + "." + getLocalBuffer().name + "[" + getLocalIndex() + ".."
-        + (getLocalIndex() + getLength()) + "[=>" + getRemoteBuffer().getVertexName() + "."
-        + getRemoteBuffer().name + "[" + getRemoteIndex() + ".." + (getRemoteIndex() + getLength()) + "[";
+    return "" + getLocalBuffer().getVertexName() + "." + getLocalBuffer().name + "[" + getLocalIndexInByte() + ".."
+        + (getLocalIndexInByte() + getLengthInByte()) + "[=>" + getRemoteBuffer().getVertexName() + "."
+        + getRemoteBuffer().name + "[" + getRemoteIndexInByte() + ".." + (getRemoteIndexInByte() + getLengthInByte())
+        + "[";
   }
 
   /**