+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/synthesis/memalloc/allocation/PiBestFitAllocator.java
@@ -186,6 +186,6 @@ public class PiBestFitAllocator extends PiOrderedAllocator {
       allocateMemoryObject(vertex, bestFitOffset);
     }
 
-    return getMemorySize();
+    return getMemorySizeInByte();
   }
 }