+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/statictools/PiMMHelper.java
@@ -370,6 +370,7 @@ public class PiMMHelper {
         graph.removeParameter((Parameter) setter);
       }
     }
+    PreesmLogger.getLogger().fine("Removing Actor: " + actor.getVertexPath());
     graph.removeActor(actor);
   }
 
@@ -410,6 +411,7 @@ public class PiMMHelper {
         if (d != null) {
           piGraph.removeDelay(d);
         }
+        PreesmLogger.getLogger().fine("Removing unused Fifo: " + f.getId());
         piGraph.removeFifo(f);
         dpi.setOutgoingFifo(null);
         dpo.setIncomingFifo(null);
@@ -453,6 +455,7 @@ public class PiMMHelper {
       if ((aa instanceof JoinActor || aa instanceof RoundBufferActor)
           && toRemoveIn.size() < aa.getDataInputPorts().size()) {
         for (DataInputPort p : toRemoveIn) {
+          PreesmLogger.getLogger().fine("Removing unused input Port: " + aa.getVertexPath() + ":" + p.getName());
           aa.getDataInputPorts().remove(p);
         }
       } else if (!toRemoveIn.isEmpty() && !(aa instanceof DelayActor)) {
@@ -463,6 +466,7 @@ public class PiMMHelper {
       if ((aa instanceof ForkActor || aa instanceof BroadcastActor)
           && toRemoveOut.size() < aa.getDataOutputPorts().size()) {
         for (DataOutputPort p : toRemoveOut) {
+          PreesmLogger.getLogger().fine("Removing unused output Port: " + aa.getVertexPath() + ":" + p.getName());
           aa.getDataOutputPorts().remove(p);
         }
       } else if (!toRemoveOut.isEmpty() && !(aa instanceof DelayActor)) {