+++ tests/org.preesm.tests.model/src/org/preesm/tests/model/scenario/types/DefaultTypeSizesTest.java
@@ -55,13 +55,31 @@ public class DefaultTypeSizesTest {
   @Test
   public void testCharTypeSize() {
     final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("char");
-    assertEquals(1L, defaultTypeSize);
+    assertEquals(8L, defaultTypeSize);
+  }
+
+  @Test
+  public void testShortTypeSize() {
+    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("short");
+    assertEquals(16L, defaultTypeSize);
   }
 
   @Test
   public void testINT32TTypeSize() {
     final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("int32_t");
-    assertEquals(4L, defaultTypeSize);
+    assertEquals(32L, defaultTypeSize);
+  }
+
+  @Test
+  public void testDoubleTypeSize() {
+    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("double");
+    assertEquals(64L, defaultTypeSize);
+  }
+
+  @Test
+  public void testLongDoubleTypeSize() {
+    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("long double");
+    assertEquals(128L, defaultTypeSize);
   }
 
 }